fix(reposerver): address PR review feedback for convoy deadlock fix
- Move convoy tests from lock_convoy_test.go into lock_test.go
- Add early context cancellation check at start of Lock()
